+++ src/template.py
@@ -95,37 +95,22 @@ in_ports = [
 hook(
     #AutoRestart(), # Use when debug directly in script.py
     OSCInterface(),
+    MemorizeScene("memorize_scene.txt")
 )
 
 #-----------------------------------------------------------------------------------------------------------
 # Class and function body
-# functions.py
-__LOGIC__
-
-#-----------------------------------------------------------------------------------------------------------
-# Filters body
-# filters.py
-#-----------------------------------------------------------------------------------------------------------
-__FILTERS__
-
-#-----------------------------------------------------------------------------------------------------------
-# Devices body
-__DEVICES__
+__MODULES__
 
 #-----------------------------------------------------------------------------------------------------------
 # Patches body
-# patches.py
+# patches/*.py
 #-----------------------------------------------------------------------------------------------------------
 __PATCHES__
 
 #-----------------------------------------------------------------------------------------------------------
-# Control body
-# control.py
-__CONTROL__
-#-----------------------------------------------------------------------------------------------------------
-
-#-----------------------------------------------------------------------------------------------------------
 # Scenes body
+# scenes/*.py
 #-----------------------------------------------------------------------------------------------------------
 _scenes = {
 __SCENES__
@@ -144,7 +129,7 @@ post = Pass()
 #post = Print('output',portnames='out')
 
 run(
-    control=_control,
+    control=control_patch,
     scenes=_scenes,
     pre=pre,
     post=post,
